# src/main/java/com/example/openapiproject/model/Shelf.java
package com.example.openapiproject.model;


import jakarta.persistence.*;
import java.util.List;

@Entity
@Table(name = "shelves")
public class Shelf {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
    private String name;
    public Shelf(Long id, String name) {
        this.id = id;
        this.name = name;
    }

    public Shelf() {
    }

    public Long getId() {
        return id;
    }
    public void setId(Long id) {
        this.id = id;
    }
    public String getName() {
        return name;
    }
    public void setName(String name) {
        this.name = name;
    }
}